uni2cube cleanup
git-svn-id: https://svn.code.sf.net/p/sauerbraten/code/src@3902 07af1644-1164-4067-b897-63bf4dc9b748

+++ shared/tools.h
@@ -1073,7 +1073,7 @@ static inline int cube2uni(uchar c)
     extern const int cube2unichars[256]; 
     return cube2unichars[c]; 
 }
-static inline int uni2cube(int c)
+static inline uchar uni2cube(int c)
 {
     extern const int uni2cubeoffsets[8];
     extern const uchar uni2cubechars[];
